requirements.txt
azure-ai-projects>=1.0.0b7
azure-ai-agents>=1.0.0
azure-identity>=1.19.0
azure-search-documents>=11.6.0
openai>=1.58.0
python-dotenv>=1.0.1
pytest>=8.3.0
ruff>=0.8.0
streamlit>=1.41.0
google-genai>=1.0.0
numpy>=1.26.0
pypdf>=4.0.0
